code/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.13.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.2.0
html5lib==1.1
idna==3.8
Jinja2==3.1.3
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
networkx==3.2.1
numpy==1.26.3
pandas==2.2.2
peewee==3.17.6
pillow==10.2.0
platformdirs==4.3.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.6
sympy==1.12
threadpoolctl==3.5.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
twilio==9.3.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yarl==1.10.0
yfinance==0.2.43
